third_party/WebKit/LayoutTests/accessibility/contenteditable-selection-expected.txt

This tests that text selection is reported correctly for contenteditable elements.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS textboxAccessible.selectionStart is 0
PASS textboxAccessible.selectionEnd is 8
PASS textboxAccessible.selectionStartLineNumber is 0
PASS textboxAccessible.selectionEndLineNumber is 1
PASS contenteditableAccessible.selectionStart is 0
PASS contenteditableAccessible.selectionEnd is 21
PASS contenteditableAccessible.selectionStartLineNumber is 0
PASS contenteditableAccessible.selectionEndLineNumber is 2
PASS textboxAccessible.selectionStart is 0
PASS textboxAccessible.selectionEnd is 6
PASS textboxAccessible.selectionStartLineNumber is 0
PASS textboxAccessible.selectionEndLineNumber is 0
PASS textboxAccessible.selectionStart is 0
PASS textboxAccessible.selectionEnd is 0
PASS line2Accessible.selectionStart is 0
PASS line2Accessible.selectionEnd is 6
PASS line2Accessible.selectionStartLineNumber is 0
PASS line2Accessible.selectionEndLineNumber is 0
PASS contenteditableAccessible.selectionStart is 0
PASS contenteditableAccessible.selectionEnd is 6
PASS contenteditableAccessible.selectionStartLineNumber is 0
PASS contenteditableAccessible.selectionEndLineNumber is 0
PASS contenteditableAccessible.selectionStart is 7
PASS contenteditableAccessible.selectionEnd is 13
PASS contenteditableAccessible.selectionStartLineNumber is 1
PASS contenteditableAccessible.selectionEndLineNumber is 1
PASS contenteditableAccessible.selectionStart is 15
PASS contenteditableAccessible.selectionEnd is 21
FAIL contenteditableAccessible.selectionStartLineNumber should be 2. Was 1.
PASS contenteditableAccessible.selectionEndLineNumber is 2
PASS contenteditableAccessible.selectionStart is 0
PASS contenteditableAccessible.selectionEnd is 6
PASS contenteditableAccessible.selectionStartLineNumber is 0
PASS contenteditableAccessible.selectionEndLineNumber is 0
PASS contenteditableAccessible.selectionStart is 7
PASS contenteditableAccessible.selectionEnd is 13
PASS contenteditableAccessible.selectionStartLineNumber is 1
PASS contenteditableAccessible.selectionEndLineNumber is 1
PASS contenteditableAccessible.selectionStart is 15
PASS contenteditableAccessible.selectionEnd is 21
FAIL contenteditableAccessible.selectionStartLineNumber should be 2. Was 1.
PASS contenteditableAccessible.selectionEndLineNumber is 2
PASS successfullyParsed is true

TEST COMPLETE
Line 1

Line 1
Line 2

Line 3
